Modifs high lvl
Répartition en trois parties :
- update des flags télécomande (system non préemptif donc inutile d update les flags de la télécomande à chaque mise à jour de l ibus)
- procces des events (erreur et changement mode de vol)
- appelle de la fonction qui correspond au mode de vol actuel
