+++ flight_controller/appli/events/flags.h
@@ -28,6 +28,7 @@ typedef enum{
 
 	FLAG_CHAN_10_PUSH,
 
+	FLAG_THROTTLE_HIGH,
 	FLAG_THROTTLE_LOW,
 	FLAG_THROTTLE_NULL,
