+++ flight_controller/appli/high_lvl/high_lvl.h
@@ -9,11 +9,16 @@
 #include "MAE.h"
 #include "settings.h"
 #include "sub/sub_action.h"
+#include "../events/events.h"
 
-#ifndef HIGH_LVL_CASES_H_
-#define HIGH_LVL_CASES_H_
+#ifndef HIGH_LVL_H_
+#define HIGH_LVL_H_
+
+//Fonctions .. bref
+void HIGH_LVL_Switch(State_drone_t * drone, State_base_t * base);
 
 
+//Fonctions pour les différents mode de vol + la fonction qui les appelle
 void HIGH_LVL_On_The_Ground(State_drone_t * drone);
 void HIGH_LVL_Manual(State_drone_t * drone);
 void HIGH_LVL_Manual_Hand_Control(State_drone_t * drone, State_base_t * base);
@@ -24,6 +29,8 @@ void HIGH_LVL_IMU_Failed_Init(State_drone_t * drone);
 void HIGH_LVL_Manual_Accro(State_drone_t * drone);
 void HIGH_LVL_Change_Pid_Settings(State_drone_t * drone);
 
+//Fonction pour la mise à jour des flags
+void HIGH_LVL_Update_Flags(State_drone_t * drone);
 
 
 #endif /* HIGH_LVL_CASES_H_ */
